# Conflicts:
#	build.gradle
#	noark-core/src/main/java/xyz/noark/core/ioc/definition/ControllerBeanDefinition.java
#	noark-core/src/main/java/xyz/noark/core/util/MathUtils.java
#	noark-core/src/test/java/xyz/noark/core/util/MathUtilsTest.java
#	noark-game/src/test/java/xyz/noark/game/ModuleControllerTest.java

+++ noark-core/src/main/java/xyz/noark/core/util/RandomUtils.java
@@ -89,7 +89,7 @@ public class RandomUtils {
 	 * @param bound 最大值（不包含）
 	 * @return 返回一个指定区间的随机数字
 	 */
-	public static long nextLong(int origin, int bound) {
+	public static long nextLong(long origin, long bound) {
 		return ThreadLocalRandom.current().nextLong(origin, bound);
 	}
